feat: open appflowy from admin web
