+++ frontend/appflowy_flutter/lib/user/application/auth/af_cloud_auth_service.dart
@@ -1,23 +1,19 @@
 import 'dart:async';
 
-import 'package:app_links/app_links.dart';
+import 'package:appflowy/startup/startup.dart';
+import 'package:appflowy/startup/tasks/appflowy_cloud_task.dart';
 import 'package:appflowy/user/application/auth/backend_auth_service.dart';
 import 'package:appflowy/user/application/auth/auth_service.dart';
 import 'package:appflowy/user/application/user_service.dart';
 import 'package:appflowy_backend/dispatch/dispatch.dart';
-import 'package:appflowy_backend/log.dart';
 import 'package:appflowy_backend/protobuf/flowy-error/errors.pb.dart';
 import 'package:appflowy_backend/protobuf/flowy-user/protobuf.dart';
 import 'package:dartz/dartz.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'auth_error.dart';
-import 'device_id.dart';
 
 class AppFlowyCloudAuthService implements AuthService {
-  final _appLinks = AppLinks();
-  StreamSubscription<Uri?>? _deeplinkSubscription;
-
   AppFlowyCloudAuthService();
 
   final BackendAuthService _backendAuthService = BackendAuthService(
@@ -66,20 +62,15 @@ class AppFlowyCloudAuthService implements AuthService {
         );
 
         final completer = Completer<Either<FlowyError, UserProfilePB>>();
-        _deeplinkSubscription = _appLinks.uriLinkStream.listen(
-          (Uri? uri) async {
-            Log.info('onDeepLink: ${uri.toString()}');
-            await _handleUri(uri, completer);
-          },
-          onError: (Object err, StackTrace stackTrace) {
-            Log.error('onDeepLinkError: ${err.toString()}', stackTrace);
-            _deeplinkSubscription?.cancel();
-            completer.complete(left(AuthError.deeplinkError));
-          },
-        );
-
-        if (!isSuccess) {
-          _deeplinkSubscription?.cancel();
+        if (isSuccess) {
+          // The [AppFlowyCloudDeepLink] must be registered before using the
+          // [AppFlowyCloudAuthService].
+          if (getIt.isRegistered<AppFlowyCloudDeepLink>()) {
+            getIt<AppFlowyCloudDeepLink>().resigerCompleter(completer);
+          } else {
+            throw Exception('AppFlowyCloudDeepLink is not registered');
+          }
+        } else {
           completer.complete(left(AuthError.signInWithOauthError));
         }
 
@@ -87,37 +78,6 @@ class AppFlowyCloudAuthService implements AuthService {
       },
       (r) => left(r),
     );
-  }
-
-  Future<void> _handleUri(
-    Uri? uri,
-    Completer<Either<FlowyError, UserProfilePB>> completer,
-  ) async {
-    if (uri != null) {
-      if (_isAuthCallbackDeeplink(uri)) {
-        // Sign in with url
-        final deviceId = await getDeviceId();
-        final payload = OauthSignInPB(
-          authType: AuthTypePB.AFCloud,
-          map: {
-            AuthServiceMapKeys.signInURL: uri.toString(),
-            AuthServiceMapKeys.deviceId: deviceId,
-          },
-        );
-        final result = await UserEventOauthSignIn(payload)
-            .send()
-            .then((value) => value.swap());
-        _deeplinkSubscription?.cancel();
-        completer.complete(result);
-      } else {
-        Log.error('onDeepLinkError: Unexpect deep link: ${uri.toString()}');
-        completer.complete(left(AuthError.signInWithOauthError));
-      }
-    } else {
-      Log.error('onDeepLinkError: Unexpect empty deep link callback');
-      _deeplinkSubscription?.cancel();
-      completer.complete(left(AuthError.emptyDeeplink));
-    }
   }
 
   @override
@@ -159,8 +119,4 @@ extension ProviderTypePBExtension on ProviderTypePB {
         throw UnimplementedError();
     }
   }
-}
-
-bool _isAuthCallbackDeeplink(Uri uri) {
-  return (uri.fragment.contains('access_token'));
 }
